BUGFIX: - Trapsheet still got removed

+++ core/report/base.py
@@ -274,7 +274,7 @@ class PaginationReport(Report):
                     file=discord.File(env.filename,
                                       filename=os.path.basename(env.filename)) if env.filename else None)
             finally:
-                if env.filename and os.path.exists(env.filename):
+                if not self.keep_image and env.filename and os.path.exists(env.filename):
                     os.remove(env.filename)
                     env.filename = None
             await view.wait()
